decouple the one-to-one relationship between scheduled runs and job IDs
this helps us avoid creating one scheduler entry per scheduled job and instead deduplicate the run IDs, increasing reliability.

+++ src/common.py
@@ -10,7 +10,13 @@ def wrap_delete_scheduled_run(cid, aid, run_id):
     sched_run = db.get_scheduled_run(cid, aid, run_id)
     if sched_run is None:
         raise Exception("Cannot find scheduled run")
-    sched_api.delete_scheduled_run(sched_run["scheduled_run_id"])
+    scheduler_id = sched_run["scheduled_run_id"]
+    sched_api.delete_scheduled_run(scheduler_id)
+    # multiple scheduled runs can have the same run ID
+    # make sure we only call delete when no other job references it
+    referenced_jobs = db.find_referenced_scheduled_runs(scheduler_id)
+    if referenced_jobs > 0:
+        return True
     if not db.delete_scheduled_run(cid, aid, run_id):
         raise Exception("Failed to delete scheduled run.")
     return True
